requirements.txt
aiogram==3.6.0
selenium==4.25.0
webdriver-manager==4.0.2
python-dotenv==1.0.1
uvloop==0.20.0

